pyproject.toml
[project]
name = "python-pw-03"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pyyaml>=6.0.3",
]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "pytest-html>=4.2.0",
    "pytest-playwright>=0.7.2",
    "python-dotenv>=1.2.2",
]
